requirements.txt
nextcord[speed]
requests
beautifulsoup4
pillow